@@ -9,9 +9,9 @@
   <url>https://korap.ids-mannheim.de/</url>
   
   <properties>
-    <java.version>1.8</java.version>
-    <jetty.version>9.4.50.v20221201</jetty.version>
-    <jersey.version>2.35</jersey.version>
+    <java.version>11</java.version>
+    <jetty.version>11.0.12</jetty.version>
+    <jersey.version>3.0.4</jersey.version>
     <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
     <project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
   </properties>
@@ -65,7 +65,7 @@
     </dependency>
     <dependency>
       <groupId>org.glassfish.jersey.containers</groupId>
-      <artifactId>jersey-container-servlet</artifactId>
+      <artifactId>jersey-container-servlet-core</artifactId>
       <version>${jersey.version}</version>
     </dependency>
 
