@@ -12,7 +12,7 @@ import java.util.Properties;
 
 import org.junit.Test;
 
-import javax.ws.rs.core.Response;
+import jakarta.ws.rs.core.Response;
 
 public class RtfExporterTest {
     
