@@ -16,6 +16,8 @@ import org.glassfish.jersey.servlet.ServletContainer;
 
 import org.tinylog.Logger;
 
+import jakarta.servlet.Servlet;
+
 /**
  * Server to provide the export web service
  */
@@ -51,7 +53,7 @@ public class PluginServer {
 
 
         ServletContainer servletContainer = new ServletContainer();        
-        ServletHolder servletHolder = new ServletHolder(servletContainer);
+        ServletHolder servletHolder = new ServletHolder((Servlet) servletContainer);
         contextHandler.addServlet(servletHolder, "/*");
         
         servletHolder.setInitOrder(0);
